+++ packages/playground/preserve-symlinks/vite.config.js
@@ -1,0 +1,6 @@
+// https://vitejs.dev/config/
+module.exports = {
+  resolve: {
+    preserveSymlinks: false
+  }
+}