+++ packages/vite/src/node/config.ts
@@ -40,6 +40,7 @@ import {
 } from './server/pluginContainer'
 import aliasPlugin from '@rollup/plugin-alias'
 import { build } from 'esbuild'
+import { performance } from 'perf_hooks'
 
 const debug = createDebugger('vite:config')
 
@@ -127,7 +128,7 @@ export interface UserConfig {
    */
   esbuild?: ESBuildOptions | false
   /**
-   * Specify additional files to be treated as static assets.
+   * Specify additional picomatch patterns to be treated as static assets.
    */
   assetsInclude?: string | RegExp | (string | RegExp)[]
   /**
@@ -285,7 +286,15 @@ export async function resolveConfig(
 
   // resolve plugins
   const rawUserPlugins = (config.plugins || []).flat().filter((p) => {
-    return p && (!p.apply || p.apply === command)
+    if (!p) {
+      return false
+    } else if (!p.apply) {
+      return true
+    } else if (typeof p.apply === 'function') {
+      return p.apply({ ...config, mode }, configEnv)
+    } else {
+      return p.apply === command
+    }
   }) as Plugin[]
   const [prePlugins, normalPlugins, postPlugins] =
     sortUserPlugins(rawUserPlugins)
@@ -440,6 +449,7 @@ export async function resolveConfig(
       ...config.optimizeDeps,
       esbuildOptions: {
         keepNames: config.optimizeDeps?.keepNames,
+        preserveSymlinks: config.resolve?.preserveSymlinks,
         ...config.optimizeDeps?.esbuildOptions
       }
     }
@@ -732,7 +742,8 @@ export async function loadConfigFromFile(
   config: UserConfig
   dependencies: string[]
 } | null> {
-  const start = Date.now()
+  const start = performance.now()
+  const getTime = () => `${(performance.now() - start).toFixed(2)}ms`
 
   let resolvedPath: string | undefined
   let isTS = false
@@ -797,17 +808,14 @@ export async function loadConfigFromFile(
         userConfig = (await eval(`import(fileUrl + '.js?t=${Date.now()}')`))
           .default
         fs.unlinkSync(resolvedPath + '.js')
-        debug(
-          `TS + native esm config loaded in ${Date.now() - start}ms`,
-          fileUrl
-        )
+        debug(`TS + native esm config loaded in ${getTime()}`, fileUrl)
       } else {
         // using eval to avoid this from being compiled away by TS/Rollup
         // append a query so that we force reload fresh config in case of
         // server restart
         userConfig = (await eval(`import(fileUrl + '?t=${Date.now()}')`))
           .default
-        debug(`native esm config loaded in ${Date.now() - start}ms`, fileUrl)
+        debug(`native esm config loaded in ${getTime()}`, fileUrl)
       }
     }
 
@@ -817,7 +825,7 @@ export async function loadConfigFromFile(
         // clear cache in case of server restart
         delete require.cache[require.resolve(resolvedPath)]
         userConfig = require(resolvedPath)
-        debug(`cjs config loaded in ${Date.now() - start}ms`)
+        debug(`cjs config loaded in ${getTime()}`)
       } catch (e) {
         const ignored = new RegExp(
           [
@@ -843,7 +851,7 @@ export async function loadConfigFromFile(
       const bundled = await bundleConfigFile(resolvedPath)
       dependencies = bundled.dependencies
       userConfig = await loadConfigFromBundledFile(resolvedPath, bundled.code)
-      debug(`bundled config file loaded in ${Date.now() - start}ms`)
+      debug(`bundled config file loaded in ${getTime()}`)
     }
 
     const config = await (typeof userConfig === 'function'