+++ packages/vite/src/node/server/hmr.ts
@@ -10,6 +10,7 @@ import { RollupError } from 'rollup'
 import match from 'minimatch'
 import { Server } from 'http'
 import { isCSSRequest } from '../plugins/css'
+import { performance } from 'perf_hooks'
 
 export const debugHmr = createDebugger('vite:hmr')
 
@@ -91,6 +92,7 @@ export async function handleHMRUpdate(
       const filteredModules = await plugin.handleHotUpdate(hmrContext)
       if (filteredModules) {
         hmrContext.modules = filteredModules
+        break
       }
     }
   }
@@ -459,17 +461,17 @@ async function readModifiedFile(file: string): Promise<string> {
 
 async function restartServer(server: ViteDevServer) {
   // @ts-ignore
-  global.__vite_start_time = Date.now()
+  global.__vite_start_time = performance.now()
   const { port } = server.config.server
-  
+
   await server.close()
-  
+
   let newServer = null
   try {
     newServer = await createServer(server.config.inlineConfig)
-  } catch (err) {
+  } catch (err: any) {
     server.config.logger.error(err.message, {
-      timestamp: true,
+      timestamp: true
     })
     return
   }