+++ packages/vite/src/node/server/index.ts
@@ -15,12 +15,12 @@ import { FSWatcher, WatchOptions } from 'types/chokidar'
 import { createWebSocketServer, WebSocketServer } from './ws'
 import { baseMiddleware } from './middlewares/base'
 import { proxyMiddleware, ProxyOptions } from './middlewares/proxy'
+import { spaFallbackMiddleware } from './middlewares/spaFallback'
 import { transformMiddleware } from './middlewares/transform'
 import {
   createDevHtmlTransformFn,
   indexHtmlMiddleware
 } from './middlewares/indexHtml'
-import history from 'connect-history-api-fallback'
 import {
   serveRawFsMiddleware,
   servePublicMiddleware,
@@ -29,7 +29,7 @@ import {
 import { timeMiddleware } from './middlewares/time'
 import { ModuleGraph, ModuleNode } from './moduleGraph'
 import { Connect } from 'types/connect'
-import { createDebugger, ensureLeadingSlash, normalizePath } from '../utils'
+import { ensureLeadingSlash, normalizePath } from '../utils'
 import { errorMiddleware, prepareError } from './middlewares/error'
 import { handleHMRUpdate, HmrOptions, handleFileAddUnlink } from './hmr'
 import { openBrowser } from './openBrowser'
@@ -53,6 +53,7 @@ import { printServerUrls } from '../logger'
 import { resolveHostname } from '../utils'
 import { searchForWorkspaceRoot } from './searchRoot'
 import { CLIENT_DIR } from '../constants'
+import { performance } from 'perf_hooks'
 
 export interface ServerOptions {
   host?: string | boolean
@@ -139,7 +140,7 @@ export interface FileSystemServeOptions {
    * Set to `false` to disable the warning
    * Default to false at this moment, will enabled by default in the future versions.
    *
-   * @expiremental
+   * @experimental
    * @default undefined
    */
   strict?: boolean | undefined
@@ -150,7 +151,7 @@ export interface FileSystemServeOptions {
    * Accepts absolute path or a path relative to project root.
    * Will try to search up for workspace root by default.
    *
-   * @expiremental
+   * @experimental
    */
   allow?: string[]
 }
@@ -237,6 +238,8 @@ export interface ViteDevServer {
   /**
    * Util for transforming a file with esbuild.
    * Can be useful for certain plugins.
+   *
+   * @deprecated import `transformWithEsbuild` from `vite` instead
    */
   transformWithEsbuild(
     code: string,
@@ -338,7 +341,7 @@ export async function createServer(
   let exitProcess: () => void
 
   const server: ViteDevServer = {
-    config: config,
+    config,
     middlewares,
     get app() {
       config.logger.warn(
@@ -502,25 +505,7 @@ export async function createServer(
 
   // spa fallback
   if (!middlewareMode || middlewareMode === 'html') {
-    middlewares.use(
-      history({
-        logger: createDebugger('vite:spa-fallback'),
-        // support /dir/ without explicit index.html
-        rewrites: [
-          {
-            from: /\/$/,
-            to({ parsedUrl }: any) {
-              const rewritten = parsedUrl.pathname + 'index.html'
-              if (fs.existsSync(path.join(root, rewritten))) {
-                return rewritten
-              } else {
-                return `/index.html`
-              }
-            }
-          }
-        ]
-      })
-    )
+    middlewares.use(spaFallbackMiddleware(root))
   }
 
   // run post config hooks
@@ -618,8 +603,10 @@ async function startServer(
   if (global.__vite_start_time) {
     info(
       chalk.cyan(
-        // @ts-ignore
-        `\n  ready in ${Date.now() - global.__vite_start_time}ms.\n`
+        `\n  ready in ${Math.round(
+          // @ts-ignore
+          performance.now() - global.__vite_start_time
+        )}ms.\n`
       )
     )
   }