+++ packages/vite/src/node/utils.ts
@@ -19,6 +19,7 @@ import {
   DecodedSourceMap,
   RawSourceMap
 } from '@ampproject/remapping/dist/types/types'
+import { performance } from 'perf_hooks'
 
 export function slash(p: string): string {
   return p.replace(/\\/g, '/')
@@ -50,12 +51,17 @@ try {
 
 const ssrExtensions = ['.js', '.cjs', '.json', '.node']
 
-export function resolveFrom(id: string, basedir: string, ssr = false): string {
+export function resolveFrom(
+  id: string,
+  basedir: string,
+  preserveSymlinks = false,
+  ssr = false
+): string {
   return resolve.sync(id, {
     basedir,
     extensions: ssr ? ssrExtensions : DEFAULT_EXTENSIONS,
     // necessary to work with pnpm
-    preserveSymlinks: isRunningWithYarnPnp || false
+    preserveSymlinks: preserveSymlinks || isRunningWithYarnPnp || false
   })
 }
 
@@ -63,11 +69,15 @@ export function resolveFrom(id: string, basedir: string, ssr = false): string {
  * like `resolveFrom` but supports resolving `>` path in `id`,
  * for example: `foo > bar > baz`
  */
-export function nestedResolveFrom(id: string, basedir: string): string {
+export function nestedResolveFrom(
+  id: string,
+  basedir: string,
+  preserveSymlinks = false
+): string {
   const pkgs = id.split('>').map((pkg) => pkg.trim())
   try {
     for (const pkg of pkgs) {
-      basedir = resolveFrom(pkg, basedir)
+      basedir = resolveFrom(pkg, basedir, preserveSymlinks)
     }
   } catch {}
   return basedir
@@ -202,8 +212,8 @@ export async function asyncReplace(
 }
 
 export function timeFrom(start: number, subtract = 0): string {
-  const time: number | string = Date.now() - start - subtract
-  const timeString = (time + `ms`).padEnd(5, ' ')
+  const time: number | string = performance.now() - start - subtract
+  const timeString = (time.toFixed(2) + `ms`).padEnd(5, ' ')
   if (time < 10) {
     return chalk.green(timeString)
   } else if (time < 50) {
@@ -566,6 +576,10 @@ export function isTargetNode(target: string | false | string[]): boolean {
 
 export function arraify<T>(target: T | T[]): T[] {
   return Array.isArray(target) ? target : [target]
+}
+
+export function toUpperCaseDriveLetter(pathName: string): string {
+  return pathName.replace(/^\w:/, (letter) => letter.toUpperCase())
 }
 
 export const multilineCommentsRE = /\/\*(.|[\r\n])*?\*\//gm